chore(cli): move commands into their own directory
