+++ packages/cli/.gitignore
@@ -26,4 +26,4 @@ write-test
 prisma-cli*.tgz
 pnpm-lock.yaml
 
-/src/__tests__/version-test-engines
+version-test-engines
