+++ packages/cli/src/__tests__/commands/Doctor.test.ts
@@ -1,5 +1,5 @@
-import { Doctor } from '../Doctor'
-import { consoleContext, Context } from './__helpers__/context'
+import { Doctor } from '../../commands/Doctor'
+import { consoleContext, Context } from '../__helpers__/context'
 
 const ctx = Context.new().add(consoleContext()).assemble()
