+++ packages/cli/src/commands/Doctor.ts
