+++ packages/cli/src/commands/Format.ts
@@ -11,7 +11,7 @@ import chalk from 'chalk'
 import fs from 'fs'
 import os from 'os'
 import path from 'path'
-import { formatms } from './utils/formatms'
+import { formatms } from '../utils/formatms'
 
 /**
  * $ prisma format
