+++ packages/cli/src/commands/Generate.ts
