+++ packages/cli/src/commands/Studio.ts
@@ -13,7 +13,7 @@ import getPort from 'get-port'
 import open from 'open'
 import path from 'path'
 
-const packageJson = require('../package.json') // eslint-disable-line @typescript-eslint/no-var-requires
+const packageJson = require('../../package.json') // eslint-disable-line @typescript-eslint/no-var-requires
 
 export class Studio implements Command {
   public instance?: StudioServer
@@ -109,7 +109,7 @@ ${chalk.bold('Examples')}
       args['--port'] || (await getPort({ port: getPort.makeRange(5555, 5600) }))
     const browser = args['--browser'] || process.env.BROWSER
 
-    const staticAssetDir = path.resolve(__dirname, '../build/public')
+    const staticAssetDir = path.resolve(__dirname, '../../build/public')
 
     const studio = new StudioServer({
       schemaPath,
@@ -120,7 +120,7 @@ ${chalk.bold('Examples')}
         resolve: {
           '@prisma/client': path.resolve(
             __dirname,
-            '../prisma-client/index.js',
+            '../../prisma-client/index.js',
           ),
           '@prisma/engines': require.resolve('@prisma/engines'),
         },
