+++ packages/cli/src/commands/Validate.ts
