+++ packages/cli/src/commands/Version.ts
