+++ packages/cli/src/utils/detectPrisma1.ts
