+++ packages/cli/src/utils/prompt/utils/deepExtend.ts
