+++ packages/cli/src/utils/prompt/utils/helpers.ts
