+++ packages/cli/src/utils/prompt/utils/isDirEmpty.ts
